init project with vite and zustand
Zustand is basically useState but with more features:

useState works really well for small applications, but useState
is local to a component.
To pass the value to a different component, we must pass it via props.

But if we need to access specific states that are scattered
around the entire application, useState doesn't really work
well.

React has a native way of doing this using the context api, but the
solution isn't really good either.

That's where Justand comes in, it allows components to define and access
global states

Signed-off-by: John Titor <50095635+JohnRTitor@users.noreply.github.com>